# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/advertisers/{advertiserId}/campaigns": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaigns list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Campaign"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CampaignCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
        "/advertisers/{advertiserId}/campaigns/{campaignId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaign by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Update campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CampaignCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaigns"
                ],
                "summary": "Delete campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
        "/advertisers/{advertiserId}/campaigns/{campaignId}/image": {
            "put": {
                "description": "Only .jpg and .png files up to 5 MB are allowed. This method won't fail if the campaign already has an image.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Upload image to campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "Fails if the campaign does not have an image",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Delete image from campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "advertiserId",
                        "name": "advertiserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "campaignId",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
...
        "/clients/bulk": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Upsert many clients at once",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Client"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Client"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
        "/clients/{clientId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Get client by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ginerr.ErrorResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ginerr.ErrorResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.AiModerationResult": {
            "type": "object",
            "properties": {
                "acceptable": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "model.Campaign": {
            "type": "object",
            "required": [
                "ad_text",
                "ad_title",
                "clicks_limit",
                "cost_per_click",
                "cost_per_impression",
                "end_date",
                "impressions_limit",
                "start_date"
            ],
            "properties": {
                "ad_text": {
                    "type": "string"
                },
                "ad_title": {
                    "type": "string"
                },
                "advertiser_id": {
                    "type": "string"
                },
                "campaign_id": {
                    "type": "string"
                },
                "clicks_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "cost_per_click": {
                    "type": "number",
                    "minimum": 0
                },
                "cost_per_impression": {
                    "type": "number",
                    "minimum": 0
                },
                "created_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "integer",
                    "minimum": 0
                },
                "image_path": {
                    "type": "string"
                },
                "impressions_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "moderation_result": {
                    "$ref": "#/definitions/model.AiModerationResult"
                },
                "start_date": {
                    "type": "integer",
                    "minimum": 0
                },
                "targeting": {
                    "$ref": "#/definitions/model.CampaignTargeting"
                }
            }
        },
        "model.CampaignCreateRequest": {
            "type": "object",
            "required": [
                "ad_text",
                "ad_title",
                "clicks_limit",
                "cost_per_click",
                "cost_per_impression",
                "end_date",
                "impressions_limit",
                "start_date"
            ],
            "properties": {
                "ad_text": {
                    "type": "string"
                },
                "ad_title": {
                    "type": "string"
                },
                "clicks_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "cost_per_click": {
                    "type": "number",
                    "minimum": 0
                },
                "cost_per_impression": {
                    "type": "number",
                    "minimum": 0
                },
                "end_date": {
                    "type": "integer",
                    "minimum": 0
                },
                "impressions_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "start_date": {
                    "type": "integer",
                    "minimum": 0
                },
                "targeting": {
                    "$ref": "#/definitions/model.CampaignTargeting"
                }
            }
        },
...
        "model.CampaignTargeting": {
            "type": "object",
            "properties": {
                "age_from": {
                    "type": "integer"
                },
                "age_to": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE",
                        "ALL"
                    ]
                },
                "location": {
                    "type": "string"
                }
            }
        },
        "model.Client": {
            "type": "object",
            "required": [
                "age",
                "client_id",
                "gender",
                "location",
                "login"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "minimum": 0
                },
                "client_id": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE"
                    ]
                },
                "location": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                }
            }
        },
...
    }
}